refactored colors function to lib
